Separate server listing and server dashboard components
