+++ src/pages/servers-page/components/listing/NewServerModalForm.js
@@ -1,13 +1,13 @@
 import React from "react";
 import {Button, Form, Modal} from "react-bootstrap";
 import PropTypes from "prop-types";
-import {BaseForm} from "../../components/BaseForm";
+import {BaseForm} from "../../../components/BaseForm";
 
 export function NewServerForm(props) {
     return (
         <Modal
             show={props.show}
-            onHide={() => {props.setShow(false)}}
+            onHide={props.onHide}
             backdrop={true}
         >
             <Modal.Header closeButton>
@@ -67,6 +67,6 @@ export function NewServerForm(props) {
 
 NewServerForm.propTypes = {
     onSubmit: PropTypes.func.isRequired,
-    show: PropTypes.bool,
-    setShow: PropTypes.func
+    onHide: PropTypes.func.isRequired,
+    show: PropTypes.bool.isRequired
 }
